refactor: [Payment/Reservation] 코드 리뷰 P2/P3 추가 개선 (#58)
- PortoneFeignConfig: feignLoggerLevel 빈 추가 (Logger.Level.FULL)
- ReservationServiceClient: status 타입 String → ReservationStatus 열거형으로 변경
- PaymentService: ReservationStatus 직접 비교 및 DataIntegrityViolationException 처리로 동시성 레이스 방어
- payments 테이블: partial unique index 추가 (PENDING/SUCCESS 중복 방지) — init.sql, DDL 양쪽 적용
- ReservationInternalDto: ReservationDetailResponse를 dto 패키지로 분리
