+++ backend/common-web/src/main/kotlin/com/ticketqueue/common/config/PortoneFeignConfig.kt
@@ -15,4 +15,7 @@ class PortoneFeignConfig {
             log.info("[PortOne] {}", String.format(format, *args).trim())
         }
     }
+
+    @Bean
+    fun feignLoggerLevel(): Logger.Level = Logger.Level.FULL
 }
